+++ backend/routes/companyRoutes.js
@@ -9,8 +9,8 @@ router.post("/", verifyToken, verifyRole('staff', 'teacher'), companyController.
 router.put("/:id", verifyToken, verifyRole('staff', 'teacher', 'student'), companyController.updateCompany);
 router.delete("/:id", verifyToken, verifyRole('staff', 'teacher'), companyController.deleteCompany);
 
-router.post("/:companyId/mentors", verifyToken, verifyRole('staff', 'teacher', 'student'), companyController.addMentor);
-router.put("/mentors/:id", verifyToken, verifyRole('staff', 'teacher', 'student'), companyController.updateMentor);
-router.delete("/mentors/:id", verifyToken, verifyRole('staff', 'teacher', 'student'), companyController.deleteMentor);
+router.post("/:companyId/mentors", verifyToken, verifyRole('staff', 'teacher'), companyController.addMentor);
+router.put("/mentors/:id", verifyToken, verifyRole('staff', 'teacher'), companyController.updateMentor);
+router.delete("/mentors/:id", verifyToken, verifyRole('staff', 'teacher'), companyController.deleteMentor);
 
 module.exports = router;